Rename user.ts to me.ts
